backend/requirements.txt
# --- Core API ---
fastapi
uvicorn[standard]
orjson               # opcional, respuestas más rápidas

# --- Manejo de datos (ETL) ---
pandas
pyarrow
PyYAML               # <-- para configs/leagues.yaml
# LanusStats         # <- comentar o instalar vía git si lo usás
# git+https://github.com/federicorabanos/futbol-data-visualizacion@main

# --- Matching / Utils ---
rapidfuzz
python-dotenv
tqdm                 # opcional, barras de progreso

# --- Conexión a Base de Datos ---
sqlalchemy>=2.0
psycopg2-binary
duckdb               # fallback local para dev

# --- Machine Learning ---
scikit-learn
xgboost

# --- (Opcional) Dependencias de scraping / notebooks ---
requests             # si tu ETL descarga desde web
ipython
webdriver-manager
beautifulsoup4

